- added ffmpeg command of converting image formats to jpg - added conditions of 4" and 6" screen sizes on product name and variant name

+++ app/jobs/convert_portrait_to_landscape_job.rb
@@ -12,25 +12,29 @@ class ConvertPortraitToLandscapeJob < ActiveJob::Base
         filetypeextension = asset_data&.filetypeextension
         testing_file = "#{Rails.root}/testing#{index}.#{filetypeextension}"
         if order.file_type == 'video'
-          dimensions = order&.variant_title&.downcase&.include?("8 pieces") ? "854:480" : "1280:720" || order&.product_title&.downcase&.include?("8 pieces") ? "854:480" : "1280:720"
+          dimensions = order&.variant_title&.downcase&.include?("8 piece") || order&.product_title&.downcase&.include?("8 piece") || order&.variant_title&.downcase&.include?("4\"") || order&.product_title&.downcase&.include?("4\"") ? "854:480" : "1280:720"
           if asset_data&.imageheight > asset_data&.imagewidth || (asset_data&.rotation == 90 || asset_data&.rotation == 270) 
             temp_file = "#{Rails.root}/temp#{index}.#{filetypeextension}"
-            system(`ffmpeg -i  "#{asset_path}" -filter_complex "[0:v]scale=ih*16/9:-1,boxblur=luma_radius='min(h\,w)/20':luma_power=1:chroma_radius='min(cw\,ch)/20':chroma_power=1[bg];[bg][0:v]overlay=(W-w)/2:(H-h)/2,crop=h=iw*9/16" "temp#{index}.#{filetypeextension}"`)
-            system(`ffmpeg -i "temp#{index}.#{filetypeextension}" -vf "scale=#{dimensions}:force_original_aspect_ratio=decrease,pad=#{dimensions}:(ow-iw)/2:(oh-ih)/2" "testing#{index}.#{filetypeextension}"`)
-            if (File.exists?(temp_file))
-              File.delete(temp_file) 
-            end
+            system(`ffmpeg -i  "#{asset_path}" -filter_complex "[0:v]scale=ih*16/9:-1,boxblur=luma_radius='min(h\,w)/20':luma_power=1:chroma_radius='min(cw\,ch)/20':chroma_power=1[bg];[bg][0:v]overlay=(W-w)/2:(H-h)/2,crop=h=iw*9/16" "#{temp_file}"`)
+            system(`ffmpeg -i "#{temp_file}" -vf "scale=#{dimensions}:force_original_aspect_ratio=decrease,pad=#{dimensions}:(ow-iw)/2:(oh-ih)/2" "#{testing_file}"`)
           else 
-            system(`ffmpeg -i "#{asset_path}" -vf "scale=#{dimensions}:force_original_aspect_ratio=decrease,pad=#{dimensions}:(ow-iw)/2:(oh-ih)/2" "testing#{index}.#{filetypeextension}"`)
+            system(`ffmpeg -i "#{asset_path}" -vf "scale=#{dimensions}:force_original_aspect_ratio=decrease,pad=#{dimensions}:(ow-iw)/2:(oh-ih)/2" "#{testing_file}"`)
           end
+          filename = "#{asset[index]&.filename}"
         else
-          dimensions = order&.variant_title&.downcase&.include?("8 pieces") ? "800:480" : "960:540" || order&.product_title&.downcase&.include?("8 pieces") ? "800:480" : "960:540"
-          system(`ffmpeg -i "#{asset_path}" -vf "scale=#{dimensions}" "testing#{index}.#{filetypeextension}"`)
+          temp_file = "#{Rails.root}/temp#{index}.#{filetypeextension}"
+          testing_file = "#{Rails.root}/testing#{index}.jpg"
+          dimensions = order&.variant_title&.downcase&.include?("8 piece") || order&.product_title&.downcase&.include?("8 piece") || order&.variant_title&.downcase&.include?("4\"") || order&.product_title&.downcase&.include?("4\"") ? "800:480" : "960:540"
+          system(`ffmpeg -i "#{asset_path}" -vf "scale=#{dimensions}" "#{temp_file}"`)
+          system(`ffmpeg -i "#{temp_file}"  "#{testing_file}"`)
+          name = "#{asset[index]&.filename}"
+          filename = name.split('.')[0] + ".jpg"
         end
         if (File.exists?(testing_file))
           downloaded_video = open(testing_file)
-          asset.attach(io: downloaded_video  , filename: "#{asset[index]&.filename}")
+          asset.attach(io: downloaded_video  , filename: filename)
           asset[index].purge
+          File.delete(temp_file) if File.exists?(temp_file)
           downloaded_video.close
           File.delete(testing_file) 
         end
